requirements.txt
djangorestframework==3.15.1   # django-5.0.6
django-admin-cli
pytz
tzdata
python-dotenv
pillow
psycopg2
django-bootstrap5
#
markdown
django-filter
djangorestframework-simplejwt
django-versatileimagefield
python-magic-bin==0.4.14
drf-flex-fields
django-cors-headers
#
django-allauth
exarth-rest-auth